--- cli/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/christheoreo/advent-of-code-2022/cli/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
