+++ cmd/cli/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/christheoreo/advent-of-code-2022/cmd/cli/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
